change float to int buy item

+++ src/main/java/TestsObjects/BuyItemByExistUserEditCC.java
@@ -107,7 +107,7 @@ public class BuyItemByExistUserEditCC extends setup
     public void PlaceOreder()
     {
         Feed feed = new Feed(driver);
-        float total = (float) (pricecheckout*1.1);
+        int total = (int) (pricecheckout*1.1);
         this.totalprice = Float.parseFloat(feed.GetTotalPriceCheckOut().substring(1));
         Assert.assertEquals(this.totalprice,total);
         BookingsPage bookings = new BookingsPage(driver);
