+++ src/main/java/TestsObjects/BuyItemByExistUserFromCart.java
@@ -11,9 +11,9 @@ import org.testng.annotations.Test;
 
 public class BuyItemByExistUserFromCart extends setup
 {
-    float priceitem;
-    float pricecheckout;
-    float totalprice;
+    int priceitem;
+    int pricecheckout;
+    int totalprice;
     String firstname ="Yinon";
     String lastname ="test";
     String address ="tel-aviv";
@@ -80,14 +80,14 @@ public class BuyItemByExistUserFromCart extends setup
     @Test(priority = 4,groups={"sanity-group"})
     public void SelectSize() {
         Feed feed = new Feed(driver);
-        this.priceitem = Float.parseFloat(feed.GetPriceItem().substring(1));
+        this.priceitem = (int) Float.parseFloat(feed.GetPriceItem().substring(1));
         Assert.assertTrue(feed.ItemRowDisplayed(),"item row not displayed");
         Assert.assertTrue(feed.CheckOutButtonNotClicibilety(),"checkout button clickable");
         feed.SelectSize();
         Assert.assertTrue(feed.CheckOutButtonclicibilety(), "checkout button not clickable");
         feed.ClickCheckOut();
         Assert.assertTrue(feed.CheckOutPageDispalyed(),"checkout page not displayed");
-        this.pricecheckout = Float.parseFloat(feed.GetPriceCheckOut().substring(1));
+        this.pricecheckout = (int)Float.parseFloat(feed.GetPriceCheckOut().substring(1));
         Assert.assertEquals(this.priceitem,pricecheckout);
     }
 
@@ -121,8 +121,8 @@ public class BuyItemByExistUserFromCart extends setup
     {
         Feed feed = new Feed(driver);
         Assert.assertTrue(feed.PlaceOrderButtonclicibilety(),"checkout button not clickable");
-        float total = (float) (pricecheckout*1.1);
-        this.totalprice = Float.parseFloat(feed.GetTotalPriceCheckOut().substring(1));
+        int total = (int) (pricecheckout*1.1);
+        this.totalprice = (int) Float.parseFloat(feed.GetTotalPriceCheckOut().substring(1));
         Assert.assertEquals(this.totalprice,total);
         feed.ClickPlaceOrder();
         String expectedTY = "Thank you for your order.";
@@ -156,14 +156,14 @@ public class BuyItemByExistUserFromCart extends setup
     @Test(priority = 7,groups={"sanity-group"})
     public void SelectSizenextitem() {
         Feed feed = new Feed(driver);
-        this.priceitem = Float.parseFloat(feed.GetPriceItem().substring(1));
+        this.priceitem = (int) Float.parseFloat(feed.GetPriceItem().substring(1));
         Assert.assertTrue(feed.ItemRowDisplayed(),"item row not displayed");
         Assert.assertTrue(feed.CheckOutButtonNotClicibilety(),"checkout button clickable");
         feed.SelectSize();
         Assert.assertTrue(feed.CheckOutButtonclicibilety(), "checkout button not clickable");
         feed.ClickCheckOut();
         Assert.assertTrue(feed.CheckOutPageDispalyed(),"checkout page not displayed");
-        this.pricecheckout = Float.parseFloat(feed.GetPriceCheckOut().substring(1));
+        this.pricecheckout = (int) Float.parseFloat(feed.GetPriceCheckOut().substring(1));
         Assert.assertEquals(this.priceitem,pricecheckout);
     }
 
@@ -197,8 +197,8 @@ public class BuyItemByExistUserFromCart extends setup
     {
         Feed feed = new Feed(driver);
         Assert.assertTrue(feed.PlaceOrderButtonclicibilety(),"checkout button not clickable");
-        float total = (float) (pricecheckout*1.1);
-        this.totalprice = Float.parseFloat(feed.GetTotalPriceCheckOut().substring(1));
+        int total = (int) (pricecheckout*1.1);
+        this.totalprice = (int) Float.parseFloat(feed.GetTotalPriceCheckOut().substring(1));
         Assert.assertEquals(this.totalprice,total);
         feed.ClickPlaceOrder();
         String expectedTY = "Thank you for your order.";
